+++ postDo/src/main/java/com/example/postDo/repository/AttachmentRepository.java
@@ -2,11 +2,9 @@ package com.example.postDo.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.postDo.entity.Account;
 import com.example.postDo.entity.Attachment;
 
-public interface AttachmentRepository extends JpaRepository<Attachment, Integer>{
+public interface AttachmentRepository extends JpaRepository<Attachment, Long>{
 
-//	Attachment findOne(Integer id);
 
 }
