+++ postDo/src/main/java/com/example/postDo/repository/FolderRepository.java
@@ -2,11 +2,9 @@ package com.example.postDo.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.postDo.entity.Account;
 import com.example.postDo.entity.Folder;
 
-public interface FolderRepository extends JpaRepository<Folder, Integer>{
+public interface FolderRepository extends JpaRepository<Folder, Long>{
 
-//	Folder findOne(Integer id);
 
 }
