+++ postDo/src/main/java/com/example/postDo/repository/MessageRepository.java
@@ -2,11 +2,9 @@ package com.example.postDo.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.postDo.entity.Account;
 import com.example.postDo.entity.Message;
 
-public interface MessageRepository extends JpaRepository<Message, Integer>{
+public interface MessageRepository extends JpaRepository<Message, Long>{
 
-//	Message findOne(Integer id);
 
 }
