+++ postDo/src/main/java/com/example/postDo/repository/RuleRepository.java
@@ -2,11 +2,9 @@ package com.example.postDo.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.example.postDo.entity.Account;
 import com.example.postDo.entity.Rule;
 
-public interface RuleRepository extends JpaRepository<Rule, Integer>{
+public interface RuleRepository extends JpaRepository<Rule, Long>{
 
-//	Rule findOne(Integer id);
 
 }
