+++ postDo/src/main/java/com/example/postDo/service/AttachmentServiceInterface.java
@@ -6,12 +6,12 @@ import com.example.postDo.entity.Attachment;
 
 public interface AttachmentServiceInterface {
 	
-	Attachment findOne(Integer id);
+	Attachment findOne(Long id);
 	
 	List<Attachment> findAll();
 	
 	Attachment save(Attachment attachment);
 	
-	void remove(Integer id);
+	void remove(Long id);
 
 }
